fix env and use i18n
