+++ src/messages/en.json
@@ -1,1 +1,3 @@
-{ }
+{
+	"nlc.training.ui.success": "NLC Training URL: *%s*."
+}
